Migrate subcommand clone to new agent mode
