+++ cmd/plakar/subcommands/agent/agent.go
@@ -39,6 +39,7 @@ import (
 	"github.com/PlakarKorp/plakar/rpc/check"
 	"github.com/PlakarKorp/plakar/rpc/checksum"
 	"github.com/PlakarKorp/plakar/rpc/cleanup"
+	"github.com/PlakarKorp/plakar/rpc/clone"
 	"github.com/PlakarKorp/plakar/rpc/info"
 	"github.com/PlakarKorp/plakar/rpc/locate"
 	"github.com/PlakarKorp/plakar/rpc/ls"
@@ -445,6 +446,18 @@ func (cmd *Agent) ListenAndServe(ctx *appcontext.AppContext) error {
 				var cmd struct {
 					Name       string
 					Subcommand cleanup.Cleanup
+				}
+				if err := msgpack.Unmarshal(rawRequest, &cmd); err != nil {
+					fmt.Fprintf(os.Stderr, "Failed to decode client request: %s\n", err)
+					return
+				}
+				subcommand = &cmd.Subcommand
+				repositoryLocation = cmd.Subcommand.RepositoryLocation
+				repositorySecret = cmd.Subcommand.RepositorySecret
+			case "clone":
+				var cmd struct {
+					Name       string
+					Subcommand clone.Clone
 				}
 				if err := msgpack.Unmarshal(rawRequest, &cmd); err != nil {
 					fmt.Fprintf(os.Stderr, "Failed to decode client request: %s\n", err)
