+++ cmd/plakar/subcommands/clone/clone.go
@@ -19,94 +19,30 @@ package clone
 import (
 	"flag"
 	"fmt"
-	"os"
-	"sync"
 
 	"github.com/PlakarKorp/plakar/appcontext"
 	"github.com/PlakarKorp/plakar/cmd/plakar/subcommands"
-	"github.com/PlakarKorp/plakar/objects"
 	"github.com/PlakarKorp/plakar/repository"
-	"github.com/PlakarKorp/plakar/storage"
-	"github.com/google/uuid"
+	"github.com/PlakarKorp/plakar/rpc"
+	"github.com/PlakarKorp/plakar/rpc/clone"
 )
 
 func init() {
-	subcommands.Register("clone", cmd_clone)
+	subcommands.Register2("clone", parse_cmd_clone)
 }
 
-func cmd_clone(ctx *appcontext.AppContext, repo *repository.Repository, args []string) (int, error) {
+func parse_cmd_clone(ctx *appcontext.AppContext, repo *repository.Repository, args []string) (rpc.RPC, error) {
 	flags := flag.NewFlagSet("clone", flag.ExitOnError)
 	flags.Parse(args)
 
 	if flags.NArg() != 2 || flags.Arg(0) != "to" {
 		ctx.GetLogger().Error("usage: %s to repository", flags.Name())
-		return 1, fmt.Errorf("usage: %s to repository", flags.Name())
+		return nil, fmt.Errorf("usage: %s to repository", flags.Name())
 	}
 
-	sourceStore := repo.Store()
-
-	configuration := sourceStore.Configuration()
-	configuration.RepositoryID = uuid.Must(uuid.NewRandom())
-
-	cloneStore, err := storage.Create(flags.Arg(1), configuration)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "%s: could not create repository: %s\n", flags.Arg(1), err)
-		return 1, err
-	}
-
-	packfileChecksums, err := sourceStore.GetPackfiles()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "%s: could not get packfiles list from repository: %s\n", sourceStore.Location(), err)
-		return 1, err
-	}
-
-	wg := sync.WaitGroup{}
-	for _, _packfileChecksum := range packfileChecksums {
-		wg.Add(1)
-		go func(packfileChecksum objects.Checksum) {
-			defer wg.Done()
-
-			rd, err := sourceStore.GetPackfile(packfileChecksum)
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "%s: could not get packfile from repository: %s\n", sourceStore.Location(), err)
-				return
-			}
-
-			err = cloneStore.PutPackfile(packfileChecksum, rd)
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "%s: could not put packfile to repository: %s\n", cloneStore.Location(), err)
-				return
-			}
-		}(_packfileChecksum)
-	}
-	wg.Wait()
-
-	indexesChecksums, err := sourceStore.GetStates()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "%s: could not get paclfiles list from repository: %s\n", sourceStore.Location(), err)
-		return 1, err
-	}
-
-	wg = sync.WaitGroup{}
-	for _, _indexChecksum := range indexesChecksums {
-		wg.Add(1)
-		go func(indexChecksum objects.Checksum) {
-			defer wg.Done()
-
-			data, err := sourceStore.GetState(indexChecksum)
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "%s: could not get index from repository: %s\n", sourceStore.Location(), err)
-				return
-			}
-
-			err = cloneStore.PutState(indexChecksum, data)
-			if err != nil {
-				fmt.Fprintf(os.Stderr, "%s: could not put packfile to repository: %s\n", cloneStore.Location(), err)
-				return
-			}
-		}(_indexChecksum)
-	}
-	wg.Wait()
-
-	return 0, nil
+	return &clone.Clone{
+		RepositoryLocation: repo.Location(),
+		RepositorySecret:   ctx.GetSecret(),
+		Dest:               flags.Arg(1),
+	}, nil
 }
